fix(convert): standardise to long to prevent implicit truncation
- Update binconv_decimal_to_binary and helper to accept long instead of int
- Fix potential integer overflow/truncamento bugs on 64-bit systems
- Increase conversion buffer size in main.c to 65 to accommodate full 64-bit binary strings

+++ include/binconv/convert.h
@@ -19,6 +19,6 @@
  * @param[in] size The total size of the output buffer.
  * @return int 0 on success, -1 on failure.
  */
-int binconv_decimal_to_binary(char *out, int value, size_t size);
+int binconv_decimal_to_binary(char *out, long value, size_t size);
 
 #endif
